+++ src/kernel/none/mp.c
@@ -239,7 +239,7 @@ roundr_up_ip(GEN x, long l)
   for(;;)
   {
     if (++x[--i]) break;
-    if (i == 2) { x[2] = HIGHBIT; setexpo(x, expo(x)+1); break; }
+    if (i == 2) { x[2] = (long)HIGHBIT; setexpo(x, expo(x)+1); break; }
   }
 }
 
@@ -704,7 +704,7 @@ divrr(GEN x, GEN y)
   if (r[0] == 0) e--;
   else if (r[0] == 1) { shift_right(r,r, 2,lr, 1,1); }
   else { /* possible only when rounding up to 0x2 0x0 ... */
-    r[2] = HIGHBIT; e++;
+    r[2] = (long)HIGHBIT; e++;
   }
   r[0] = evaltyp(t_REAL)|evallg(lr);
   r[1] = evalsigne(sx) | evalexpo(e);
